Remove old ReportDiff
