+++ lib/pmdtester.rb
@@ -5,13 +5,11 @@ require 'logger'
 require_relative 'pmdtester/cmd'
 require_relative 'pmdtester/pmd_branch_detail'
 require_relative 'pmdtester/pmd_configerror'
-require_relative 'pmdtester/pmd_fileinfo'
 require_relative 'pmdtester/pmd_error'
 require_relative 'pmdtester/pmd_report_detail'
 require_relative 'pmdtester/pmd_violation'
 require_relative 'pmdtester/project'
 require_relative 'pmdtester/report_diff'
-require_relative 'pmdtester/fast_report_diff'
 require_relative 'pmdtester/resource_locator'
 require_relative 'pmdtester/runner'
